+++ tests/testthat/test-api.R
@@ -1262,3 +1262,29 @@ describe("cross-device eager (check_eager)", {
     check_eager(nv_argmin, sortable)
   })
 })
+
+describe("nv_flatten", {
+  it("works for 2D input", {
+    x <- matrix(1:4, nrow = 2)
+    expect_equal(
+      nv_flatten(nv_array(x)),
+      nv_array(as.vector(t(x)))
+    )
+  })
+  it("works for 1D input", {
+    x <- nv_array(1:3)
+    expect_equal(
+      nv_flatten(x),
+      x
+    )
+  })
+  it("fails for 0D input", {
+    expect_error(nv_flatten(1), "scalar")
+  })
+  it("works with empty input", {
+    expect_equal(
+      nv_flatten(nv_empty("f32", c(2, 0))),
+      nv_empty("f32", 0L)
+    )
+  })
+})